Fix some issues with the xds-protos package (#25999)
* Fix issues with xds-protos && add a README

* Split the build process from setup.py
